Adds the mlmodels app
