+++ Pipfile
@@ -33,7 +33,6 @@ social-auth-app-django = "*"
 gunicorn = "*"
 raven = "*"
 django-celery-email = "*"
-nbconvert = "*"
 simpleitk = "*"
 django-celery-beat = "*"
 django-favicon-plus = "*"
