+++ app/grandchallenge/algorithms/templates/algorithms/algorithm_form.html
@@ -11,5 +11,5 @@
     <h2>Create An Algorithm</h2>
 
     {% crispy form %}
-
+    <input type="submit" value="Save" class="btn btn-primary"/>
 {% endblock %}
