+++ app/grandchallenge/algorithms/urls.py
@@ -1,6 +1,5 @@
 from django.urls import path
 
-from grandchallenge.algorithms.forms import algorithm_upload_widget
 from grandchallenge.algorithms.views import (
     AlgorithmList,
     AlgorithmCreate,
@@ -14,11 +13,6 @@ app_name = "algorithms"
 urlpatterns = [
     path("", AlgorithmList.as_view(), name="list"),
     path("create/", AlgorithmCreate.as_view(), name="create"),
-    path(
-        f"create/{algorithm_upload_widget.ajax_target_path}",
-        algorithm_upload_widget.handle_ajax,
-        name="algorithm-upload-ajax",
-    ),
     path("<slug:slug>/", AlgorithmDetail.as_view(), name="detail"),
     path(
         "<slug:slug>/run/",
