+++ app/grandchallenge/cases/tasks.py
@@ -7,7 +7,6 @@ from uuid import UUID
 from celery import shared_task
 from django.db import transaction
 
-from grandchallenge.algorithms.models import Job
 from grandchallenge.cases.image_builders import ImageBuilderResult
 from grandchallenge.cases.image_builders.tiff import image_builder_tiff
 from grandchallenge.cases.image_builders.metaio_mhd_mha import (
@@ -25,6 +24,7 @@ from grandchallenge.jqfileupload.widgets.uploader import (
     StagedAjaxFile,
     NotFoundError,
 )
+from grandchallenge.mlmodels.models import Job
 
 
 class ProvisioningError(Exception):
@@ -259,16 +259,14 @@ def build_images(upload_session_uuid: UUID):
                 if upload_session.annotationset:
                     upload_session.annotationset.images.add(*collected_images)
 
-                if upload_session.algorithm:
+                if upload_session.mlmodel:
                     for image in collected_images:
                         Job.objects.create(
-                            algorithm=upload_session.algorithm, image=image
+                            mlmodel=upload_session.mlmodel, image=image
                         )
 
-                if upload_session.algorithm_result:
-                    upload_session.algorithm_result.images.add(
-                        *collected_images
-                    )
+                if upload_session.mlmodel_result:
+                    upload_session.mlmodel_result.images.add(*collected_images)
 
                 # Delete any touched file data
                 for file in session_files:
